+++ morefocus-base/src/main/java/cn/morefocus/base/module/support/serialnumber/mapper/SerialNumberRecordMapper.java
@@ -13,8 +13,6 @@ import java.util.List;
 
 /**
  * 单据序列号 生成的记录
- *
- *
  */
 @Mapper
 @Component
@@ -22,30 +20,16 @@ public interface SerialNumberRecordMapper extends BaseMapper<SerialNumberRecordE
 
     /**
      * 根据 id和日期 查询 记录id
-     *
-     * @param serialNumberId
-     * @param recordDate
-     * @return
      */
     Long selectRecordIdBySerialNumberIdAndDate(@Param("serialNumberId") Integer serialNumberId, @Param("recordDate") String recordDate);
 
     /**
      * 更新记录
-     *
-     * @param serialNumberId
-     * @param recordDate
-     * @param lastNumber
-     * @param count
-     * @return
      */
-    Long updateRecord(@Param("serialNumberId") Integer serialNumberId, @Param("recordDate") LocalDate recordDate, @Param("lastNumber") Long lastNumber, @Param("count") int count);
+    Long updateRecord(@Param("serialNumberId") Long serialNumberId, @Param("recordDate") LocalDate recordDate, @Param("lastNumber") Long lastNumber, @Param("count") int count);
 
     /**
      * 分页查询记录
-     *
-     * @param page
-     * @param queryForm
-     * @return
      */
     List<SerialNumberRecordEntity> query(Page page, @Param("queryForm") SerialNumberRecordQueryForm queryForm);
 }
